kafka_prometheus_connector/requirements.txt
fastapi
uvicorn
pydantic
jinja2
pydantic-settings
kafka-python
